+++ AmazonChimeSDKDemo/AmazonChimeSDKDemo/MeetingModel.swift
@@ -23,6 +23,7 @@ class MeetingModel: NSObject {
 
     // Dependencies
     let meetingId: String
+    let meetingEndpointUrl: String
     let selfName: String
     var audioVideoConfig = AudioVideoConfiguration()
     let callKitOption: CallKitOption
@@ -124,8 +125,10 @@ class MeetingModel: NSObject {
          meetingId: String,
          selfName: String,
          audioVideoConfig: AudioVideoConfiguration,
-         callKitOption: CallKitOption) {
+         callKitOption: CallKitOption,
+         meetingEndpointUrl: String) {
         self.meetingId = meetingId
+        self.meetingEndpointUrl = meetingEndpointUrl.isEmpty ? AppConfiguration.url : meetingEndpointUrl
         self.selfName = selfName
         self.audioVideoConfig = audioVideoConfig
         self.callKitOption = callKitOption
@@ -199,8 +202,7 @@ class MeetingModel: NSObject {
     }
     
     func postStopTranscriptionRequest() {
-        var url = AppConfiguration.url
-        url = url.hasSuffix("/") ? url : "\(url)/"
+        let url = self.meetingEndpointUrl.hasSuffix("/") ? self.meetingEndpointUrl : "\(self.meetingEndpointUrl)/"
         let encodedURL = HttpUtils.encodeStrForURL(
                 str: "\(url)stop_transcription?title=\(meetingId)")
         HttpUtils.postRequest(url: encodedURL, jsonData: nil) { _, error in
@@ -228,6 +230,13 @@ class MeetingModel: NSObject {
             }
         }
         return displayName
+    }
+    
+    func getVideoTileAttendeeId(for indexPath: IndexPath) -> String {
+        if let videoTileState = videoModel.getVideoTileState(for: indexPath) {
+            return videoTileState.attendeeId
+        }
+        return ""
     }
 
     func chooseAudioDevice(_ audioDevice: MediaDevice) {
@@ -430,6 +439,22 @@ extension MeetingModel: AudioVideoObserver {
 
     func videoSessionDidStartConnecting() {
         logWithFunctionName()
+    }
+    
+    func remoteVideoSourcesDidBecomeAvailable(sources: [RemoteVideoSource]) {
+        logWithFunctionName()
+        sources.forEach { source in
+            // Initialize with defaults in case we want to update through UI
+            videoModel.remoteVideoSourceConfigurations[source] = VideoSubscriptionConfiguration()
+        }
+        // Use default auto-subscribe behavior
+    }
+    
+    func remoteVideoSourcesDidBecomeUnavailable(sources: [RemoteVideoSource]) {
+        logWithFunctionName()
+        sources.forEach { source in
+            videoModel.remoteVideoSourceConfigurations.removeValue(forKey: source)
+        }
     }
 
     func videoSessionDidStartWithStatus(sessionStatus: MeetingSessionStatus) {